fix(api): make fabnode webhook working
